Separated message sending into its own thread
Message sending is now running in its own thread. This was done so that
if a message fails to send, other messages can continue to be queued up.
However, at the moment there is nothing in place to delay sending so it
will probably get rate limited quickly if any errors occur.